feat(orientation): add new implementation for get/set orientation
