+++ src/tools/index.ts
@@ -35,8 +35,7 @@ import dragAndDrop from './interactions/drag-and-drop.js';
 import setValue from './interactions/set-value.js';
 import getText from './interactions/get-text.js';
 import getPageSource from './interactions/get-page-source.js';
-import getOrientation from './interactions/get-orientation.js';
-import setOrientation from './interactions/set-orientation.js';
+import { getOrientation, setOrientation } from './interactions/orientation.js';
 import handleAlert from './interactions/handle-alert.js';
 import { screenshot, elementScreenshot } from './interactions/screenshot.js';
 import activateApp from './app-management/activate-app.js';
